+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 
 ## [Unreleased]
 
+## [v0.3.0] - 2024-07-10
+### Added
+- array manipulation functions
+### Fixed
+- GIT export properties
+
 ## [v0.2.0] - 2024-07-04
 ### Added
 - Modules now can listen to `onEnable` and `onDisable` events.
@@ -17,5 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org).
 - SlickModuleInterface, WebModuleInterface and ConsoleModuleInterface
 - Basic implementations for Module and middleware handler.
 
-[Unreleased]: https://github.com/slickframework/module-api/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/slickframework/module-api/compare/v0.3.0...HEAD
+[v0.3.0]: https://github.com/slickframework/module-api/compare/v0.2.0...v0.3.0
+[v0.2.0]: https://github.com/slickframework/module-api/compare/v0.1.0...v0.2.0
 [v0.1.0]: https://github.com/slickframework/module-api/compare/9efd6e...v0.1.0